+++ src/navigation/types.ts
@@ -1,0 +1,13 @@
+export type AuthStackParamList = {
+  Login: undefined;
+  BiometricUnlock: undefined;
+};
+
+export type AppStackParamList = {
+  HomePlaceholder: undefined;
+};
+
+export type RootStackParamList = {
+  Auth: undefined;
+  App: undefined;
+};